catalog: Refactor transaction logic to module
This commit extracts the catalog transaction logic from the
`mz_adapter::catalog` module to the `mz_adapter::catalog::transact`
module. This is purely code movement and does not involve any code
changes.
